moved PointPOST to DTO project
