+++ Backend/DTO/DTO.csproj
@@ -52,6 +52,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DTO\PointDTO.cs" />
     <Compile Include="DTO\RecoverableErrorDTO.cs" />
     <Compile Include="DTO\LogRequestDTO.cs" />
     <Compile Include="KFCSSO_API\SSO_API_Requests.cs" />
