+++ Backend/WebApi-PointMap/Controllers/TesterController.cs
@@ -6,7 +6,6 @@ using System.Net.Http;
 using System.Text;
 using System.Web.Http;
 using System.Web.Http.Cors;
-using WebApi_PointMap.Models;
 
 namespace WebApi_PointMap.Controllers
 {
